Added login/logout change
