+++ src/components/EditChefModal/index.js
@@ -27,6 +27,8 @@ function EditChefModal(props) {
                 <TextField fullWidth onChange={props.handleInputChange} value={props.chef.first} type="text" name="first" />
                 <TextField fullWidth onChange={props.handleInputChange} value={props.chef.last} type="text" name="last" />
                 <TextField fullWidth onChange={props.handleInputChange} value={props.chef.bio} type="text" name="bio" />
+                <Button variant="contained" component="label" onChange={props.fileChange} val={props.file}>Upload<input type="file" hidden /></Button>
+				<Button className={classes.button} variant="contained" color="secondary" onClick={() => props.uploadToCloudinary(props.file)}>Add Photo</Button>	
                 <TextField fullWidth onChange={props.handleInputChange} value={props.chef.zipcode} type="text" name="zipcode" />
                 <TextField fullWidth onChange={props.handleInputChange} value={props.chef.profilePic} type="text" name="profilePic" />
                 <FormGroup row>
